+++ src/main/java/webserver/WebServer.java
@@ -2,7 +2,6 @@ package webserver;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import webserver.handler.RequestHandler;
 
 import java.net.ServerSocket;
 import java.net.Socket;
