+++ src/test/java/webserver/RequestHandlerTest.java
@@ -4,7 +4,6 @@ import db.DataBase;
 import org.junit.jupiter.api.Test;
 import support.StubSocket;
 import utils.FileIoUtils;
-import webserver.handler.RequestHandler;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
@@ -77,7 +76,7 @@ class RequestHandlerTest {
 
         var expected = "HTTP/1.1 200 OK \r\n" +
                 "Content-Type: text/html;charset=utf-8 \r\n" +
-                "Content-Length: 6902 \r\n" +
+                "Content-Length: 7122 \r\n" +
                 "\r\n" +
                 new String(Objects.requireNonNull(FileIoUtils.loadFileFromClasspath("templates/index.html")));
